+++ parallel_gca.h
@@ -3,6 +3,6 @@
 
 #include <stdbool.h>
 
-float* gca(float (*target_function)(float*, int), float lb, float ub, int dim, int global_pop_size, int n_iter, int my_rank, int local_pop_size, bool debug);
+double* gca(double (*target_function)(double*, int), double lb, double ub, int dim, int global_pop_size, int n_iter, int my_rank, int local_pop_size, bool debug);
 
 #endif
